Module rustc_mir::graphviz

Unstable (rustc_private)

Functions

write_mir_graphviz 	 [Unstable]
Write a graphviz DOT graph of a list of MIRs.

write_node_label 	 [Unstable]
Write a graphviz HTML-styled label for the given basic block, with all necessary escaping already performed. (This is suitable for emitting directly, as is done in this module, or for use with the LabelText::HtmlStr from libgraphviz.)