Trait rustc_mir::hair::Hair

pub trait Hair: Sized + Debug + Clone + Eq + Hash {
    type VarId: Copy + Debug + Eq + Hash;
    type DefId: Copy + Debug + Eq + Hash;
    type AdtDef: Copy + Debug + Eq + Hash;
    type Name: Copy + Debug + Eq + Hash;
    type Ident: Copy + Debug + Eq + Hash;
    type InternedString: Clone + Debug + Eq + Hash;
    type Bytes: Clone + Debug + Eq + Hash;
    type Span: Copy + Debug + Eq;
    type Projection: Clone + Debug + Eq;
    type Substs: Clone + Debug + Eq;
    type ClosureSubsts: Clone + Debug + Eq;
    type Ty: Clone + Debug + Eq;
    type Region: Copy + Debug;
    type CodeExtent: Copy + Debug + Hash + Eq;
    type Pattern: Clone + Debug + Mirror<Self, Output=Pattern<Self>>;
    type Expr: Clone + Debug + Mirror<Self, Output=Expr<Self>>;
    type Stmt: Clone + Debug + Mirror<Self, Output=Stmt<Self>>;
    type Block: Clone + Debug + Mirror<Self, Output=Block<Self>>;
    type InlineAsm: Clone + Debug + Eq + Hash;
    fn unit_ty(&mut self) -> Self::Ty;
    fn usize_ty(&mut self) -> Self::Ty;
    fn bool_ty(&mut self) -> Self::Ty;
    fn partial_eq(&mut self, ty: Self::Ty) -> ItemRef<Self>;
    fn partial_le(&mut self, ty: Self::Ty) -> ItemRef<Self>;
    fn num_variants(&mut self, adt: Self::AdtDef) -> usize;
    fn fields(&mut self, adt: Self::AdtDef, variant_index: usize) -> Vec<Field<Self>>;
    fn needs_drop(&mut self, ty: Self::Ty, span: Self::Span) -> bool;
    fn span_bug(&mut self, span: Self::Span, message: &str) -> !;

    fn mirror<M: Mirror<Self>>(&mut self, ast: M) -> M::Output { ... }
}

Associated Types

type VarId: Copy + Debug + Eq + Hash

type DefId: Copy + Debug + Eq + Hash

type AdtDef: Copy + Debug + Eq + Hash

type Name: Copy + Debug + Eq + Hash

type Ident: Copy + Debug + Eq + Hash

type InternedString: Clone + Debug + Eq + Hash

type Bytes: Clone + Debug + Eq + Hash

type Span: Copy + Debug + Eq

type Projection: Clone + Debug + Eq

type Substs: Clone + Debug + Eq

type ClosureSubsts: Clone + Debug + Eq

type Ty: Clone + Debug + Eq

type Region: Copy + Debug

type CodeExtent: Copy + Debug + Hash + Eq

type Pattern: Clone + Debug + Mirror<Self, Output=Pattern<Self>>

type Expr: Clone + Debug + Mirror<Self, Output=Expr<Self>>

type Stmt: Clone + Debug + Mirror<Self, Output=Stmt<Self>>

type Block: Clone + Debug + Mirror<Self, Output=Block<Self>>

type InlineAsm: Clone + Debug + Eq + Hash

Required Methods

fn unit_ty(&mut self) -> Self::Ty

Returns the unit type ()

fn usize_ty(&mut self) -> Self::Ty

Returns the type usize.

fn bool_ty(&mut self) -> Self::Ty

Returns the type bool.

fn partial_eq(&mut self, ty: Self::Ty) -> ItemRef<Self>

Returns a reference to PartialEq::<T,T>::eq

fn partial_le(&mut self, ty: Self::Ty) -> ItemRef<Self>

Returns a reference to PartialOrd::<T,T>::le

fn num_variants(&mut self, adt: Self::AdtDef) -> usize

Returns the number of variants for the given enum

fn fields(&mut self, adt: Self::AdtDef, variant_index: usize) -> Vec<Field<Self>>

fn needs_drop(&mut self, ty: Self::Ty, span: Self::Span) -> bool

true if a value of type ty (may) need to be dropped; this may return false even for non-Copy types if there is no destructor to execute. If correct result is not known, may be approximated by returning true; this will result in more drops but not incorrect code.

fn span_bug(&mut self, span: Self::Span, message: &str) -> !

Report an internal inconsistency.

Provided Methods

fn mirror<M: Mirror<Self>>(&mut self, ast: M) -> M::Output

Normalizes ast into the appropriate mirror type.