Module rustc_mir::interpret

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

An interpreter for MIR used in CTFE and by miri

Structs

CompileTimeEvaluator 	 [
Experimental
]

EvalContext 	 [
Experimental
]

Frame 	 [
Experimental
]
A stack frame.

Memory 	 [
Experimental
]

ResourceLimits 	 [
Experimental
]

TyAndPacked 	 [
Experimental
]

ValTy 	 [
Experimental
]

Enums

MemoryKind 	 [
Experimental
]

Place 	 [
Experimental
]

PlaceExtra 	 [
Experimental
]

StackPopCleanup 	 [
Experimental
]

Traits

HasMemory 	 [
Experimental
]

Machine 	 [
Experimental
]
Methods of this trait signifies a point where CTFE evaluation would fail and some use case dependent behaviour can instead be applied

Functions

const_eval_provider 	 [
Experimental
]

eval_body 	 [
Experimental
]

eval_body_as_integer 	 [
Experimental
]