Struct rustc_mir::interpret::Memory

pub struct Memory<'a, 'tcx: 'a, M: Machine<'tcx>> {
    pub data: M::MemoryData,
    pub cur_frame: usize,
    pub tcx: TyCtxt<'a, 'tcx, 'tcx>,
    // some fields omitted
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Fields
data: M::MemoryData

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Additional data required by the Machine

cur_frame: usize

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

The current stack frame. Used to check accesses against locks.

tcx: TyCtxt<'a, 'tcx, 'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Methods

impl<'a, 'tcx, M: Machine<'tcx>> Memory<'a, 'tcx, M>

pub fn new(
tcx: TyCtxt<'a, 'tcx, 'tcx>,
max_memory: u64,
data: M::MemoryData
) -> Self

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn allocations<'x>(
&'x self
) -> impl Iterator<Item = (AllocId, &'x Allocation)>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn create_fn_alloc(&mut self, instance: Instance<'tcx>) -> MemoryPointer

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn allocate_cached(&mut self, bytes: &[u8]) -> MemoryPointer

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn allocate(
&mut self,
size: u64,
align: u64,
kind: Option<MemoryKind<M::MemoryKinds>>
) -> EvalResult<'tcx, MemoryPointer>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

kind is None for statics

pub fn reallocate(
&mut self,
ptr: MemoryPointer,
old_size: u64,
old_align: u64,
new_size: u64,
new_align: u64,
kind: MemoryKind<M::MemoryKinds>
) -> EvalResult<'tcx, MemoryPointer>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn deallocate_local(&mut self, ptr: MemoryPointer) -> EvalResult<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn deallocate(
&mut self,
ptr: MemoryPointer,
size_and_align: Option<(u64, u64)>,
kind: MemoryKind<M::MemoryKinds>
) -> EvalResult<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn pointer_size(&self) -> u64

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn endianess(&self) -> Endian

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn check_align(
&self,
ptr: Pointer,
align: u64,
access: Option<AccessKind>
) -> EvalResult<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Check that the pointer is aligned AND non-NULL.

pub fn check_bounds(&self, ptr: MemoryPointer, access: bool) -> EvalResult<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

impl<'a, 'tcx, M: Machine<'tcx>> Memory<'a, 'tcx, M>

Allocation accessors

pub fn get(&self, id: AllocId) -> EvalResult<'tcx, &Allocation>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn get_fn(&self, ptr: MemoryPointer) -> EvalResult<'tcx, Instance<'tcx>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn dump_alloc(&self, id: AllocId)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

For debugging, print an allocation and all allocations it points to, recursively.

pub fn dump_allocs(&self, allocs: Vec<AllocId>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

For debugging, print a list of allocations and all allocations they point to, recursively.

pub fn leak_report(&self) -> usize

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

impl<'a, 'tcx, M: Machine<'tcx>> Memory<'a, 'tcx, M>

Reading and writing

pub fn mark_static_initalized(
&mut self,
alloc_id: AllocId,
mutability: Mutability
) -> EvalResult<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

mark an allocation as static and initialized, either mutable or not

pub fn copy(
&mut self,
src: Pointer,
dest: Pointer,
size: u64,
align: u64,
nonoverlapping: bool
) -> EvalResult<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn read_c_str(&self, ptr: MemoryPointer) -> EvalResult<'tcx, &[u8]>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn read_bytes(&self, ptr: Pointer, size: u64) -> EvalResult<'tcx, &[u8]>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn write_bytes(&mut self, ptr: Pointer, src: &[u8]) -> EvalResult<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn write_repeat(
&mut self,
ptr: Pointer,
val: u8,
count: u64
) -> EvalResult<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn read_primval(
&self,
ptr: MemoryPointer,
size: u64,
signed: bool
) -> EvalResult<'tcx, PrimVal>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn read_ptr_sized_unsigned(
&self,
ptr: MemoryPointer
) -> EvalResult<'tcx, PrimVal>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn write_primval(
&mut self,
ptr: MemoryPointer,
val: PrimVal,
size: u64,
signed: bool
) -> EvalResult<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn write_ptr_sized_unsigned(
&mut self,
ptr: MemoryPointer,
val: PrimVal
) -> EvalResult<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

impl<'a, 'tcx, M: Machine<'tcx>> Memory<'a, 'tcx, M>

Undefined bytes

pub fn mark_definedness(
&mut self,
ptr: Pointer,
size: u64,
new_state: bool
) -> EvalResult<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?