Enum rustc_mir::repr::Lvalue

pub enum Lvalue<'tcx> {
    Var(u32),
    Temp(u32),
    Arg(u32),
    Static(DefId),
    ReturnPointer,
    Projection(Box<LvalueProjection<'tcx>>),
}

A path to a value; something that can be evaluated without changing or disturbing program state.

Variants

Var

local variable declared by the user

Temp

temporary introduced during lowering into MIR

Arg

formal parameter of the function; note that these are NOT the bindings that the user declares, which are vars

Static

static or static mut variable

ReturnPointer

the return pointer of the fn

Projection

projection out of an lvalue (access a field, deref a pointer, etc)

Methods

impl<'tcx> Lvalue<'tcx>

fn field(self, f: Field) -> Lvalue<'tcx>

fn deref(self) -> Lvalue<'tcx>

fn index(self, index: Operand<'tcx>) -> Lvalue<'tcx>

fn elem(self, elem: LvalueElem<'tcx>) -> Lvalue<'tcx>