Enum rustc_mir::repr::Terminator

pub enum Terminator<H: Hair> {
    Goto {
        target: BasicBlock,
    },
    Panic {
        target: BasicBlock,
    },
    If {
        cond: Operand<H>,
        targets: [BasicBlock; 2],
    },
    Switch {
        discr: Lvalue<H>,
        targets: Vec<BasicBlock>,
    },
    Diverge,
    Return,
    Call {
        data: CallData<H>,
        targets: [BasicBlock; 2],
    },
}

Variants

Goto

block should have one successor in the graph; we jump there

Fields

target

Panic

block should initiate unwinding; should be one successor that does cleanup and branches to DIVERGE_BLOCK

Fields

target

If

jump to branch 0 if this lvalue evaluates to true

Fields

cond

targets

Switch

lvalue evaluates to some enum; jump depending on the branch

Fields

discr

targets

Diverge

Indicates that the last statement in the block panics, aborts, etc. No successors. This terminator appears on exactly one basic block which we create in advance. However, during construction, we use this value as a sentinel for "terminator not yet assigned", and assert at the end that only the well-known diverging block actually diverges.

Return

Indicates a normal return. The ReturnPointer lvalue should have been filled in by now. This should only occur in the END_BLOCK.

Call

block ends with a call; it should have two successors. The first successor indicates normal return. The second indicates unwinding.

Fields

data

targets

Methods

impl<H: Hair> Terminator<H>

fn successors(&self) -> &[BasicBlock]