Struct rustc_mir::repr::Mir

pub struct Mir<'tcx> {
    pub basic_blocks: Vec<BasicBlockData<'tcx>>,
    pub return_ty: FnOutput<'tcx>,
    pub var_decls: Vec<VarDecl<'tcx>>,
    pub arg_decls: Vec<ArgDecl<'tcx>>,
    pub temp_decls: Vec<TempDecl<'tcx>>,
}

Lowered representation of a single function.

Fields

basic_blocks

List of basic blocks. References to basic block use a newtyped index type BasicBlock that indexes into this vector.

return_ty

Return type of the function.

var_decls

Variables: these are stack slots corresponding to user variables. They may be assigned many times.

arg_decls

Args: these are stack slots corresponding to the input arguments.

temp_decls

Temp declarations: stack slots that for temporaries created by the compiler. These are assigned once, but they are not SSA values in that it is possible to borrow them and mutate them through the resulting reference.

Methods

impl<'tcx> Mir<'tcx>

fn all_basic_blocks(&self) -> Vec<BasicBlock>

fn basic_block_data(&self, bb: BasicBlock) -> &BasicBlockData<'tcx>

fn basic_block_data_mut(&mut self, bb: BasicBlock) -> &mut BasicBlockData<'tcx>

impl<'tcx> Mir<'tcx>

fn operand_ty(&self, tcx: &ctxt<'tcx>, operand: &Operand<'tcx>) -> Ty<'tcx>

fn binop_ty(&self, tcx: &ctxt<'tcx>, op: BinOp, lhs_ty: Ty<'tcx>, rhs_ty: Ty<'tcx>) -> Ty<'tcx>

fn lvalue_ty(&self, tcx: &ctxt<'tcx>, lvalue: &Lvalue<'tcx>) -> LvalueTy<'tcx>