Struct rustc_mir::transform::dump_mir::DumpMir

pub struct DumpMir;

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?