Struct rustc_mir::transform::nll::RegionIndex

pub struct RegionIndex(pub u32);

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?