Module rustc_mir::transform::promote_consts

Unstable (rustc_private)

A pass that promotes borrows of constant rvalues.

The rvalues considered constant are trees of temps, each with exactly one initialization, and holding a constant value with no interior mutability. They are placed into a new MIR constant body in promoted and the borrow rvalue is replaced with a Literal::Promoted using the index into promoted of that constant MIR.

This pass assumes that every use is dominated by an initialization and can otherwise silence errors, if move analysis runs after promotion on broken MIR.

Enums

Candidate 	 [Unstable]
A "root candidate" for promotion, which will become the returned value in a promoted MIR, unless it's a subset of a larger candidate.

TempState 	 [Unstable]
State of a temporary during collection and promotion.

Functions

collect_temps 	 [Unstable]

promote_candidates 	 [Unstable]