Module rustc_mir::transform::rustc_peek

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Structs

SanityCheck 	 [
Experimental
]

Functions

sanity_check_via_rustc_peek 	 [
Experimental
]
This function scans mir for all calls to the intrinsic rustc_peek that have the expression form rustc_peek(&expr).