Struct rustc_mir::transform::type_check::MirTypeckRegionConstraints

pub struct MirTypeckRegionConstraints<'tcx> {
    pub liveness_set: Vec<(Region<'tcx>, Location)>,
    pub outlives_sets: Vec<OutlivesSet<'tcx>>,
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

A collection of region constraints that must be satisfied for the program to be considered well-typed.

Fields
liveness_set: Vec<(Region<'tcx>, Location)>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

In general, the type-checker is not responsible for enforcing liveness constraints; this job falls to the region inferencer, which performs a liveness analysis. However, in some limited cases, the MIR type-checker creates temporary regions that do not otherwise appear in the MIR -- in particular, the late-bound regions that it instantiates at call-sites -- and hence it must report on their liveness constraints.

outlives_sets: Vec<OutlivesSet<'tcx>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

During the course of type-checking, we will accumulate region constraints due to performing subtyping operations or solving traits. These are accumulated into this vector for later use.