Module rustc_mir::util

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Modules

borrowck_errors 	 [
Experimental
]

def_use 	 [
Experimental
]
Def-use analysis.

elaborate_drops 	 [
Experimental
]

liveness 	 [
Experimental
]
Liveness analysis which computes liveness of MIR local variables at the boundary of basic blocks

patch 	 [
Experimental
]

Enums

PassWhere 	 [
Experimental
]
An indication of where we are in the control flow graph. Used for printing extra information in dump_mir

Functions

dump_enabled 	 [
Experimental
]

dump_mir 	 [
Experimental
]
If the session is properly configured, dumps a human-readable representation of the mir into:

write_graphviz_node_label 	 [
Experimental
]
Write a graphviz HTML-styled label for the given basic block, with all necessary escaping already performed. (This is suitable for emitting directly, as is done in this module, or for use with the LabelText::HtmlStr from libgraphviz.)

write_mir_graphviz 	 [
Experimental
]
Write a graphviz DOT graph of a list of MIRs.

write_mir_pretty 	 [
Experimental
]
Write out a human-readable textual representation for the given MIR.