Trait rustc_mir::util::borrowck_errors::BorrowckErrors

pub trait BorrowckErrors {
    fn struct_span_err_with_code<'a, S: Into<MultiSpan>>(
        &'a self, 
        sp: S, 
        msg: &str, 
        code: &str
    ) -> DiagnosticBuilder<'a>;

    fn struct_span_err<'a, S: Into<MultiSpan>>(
        &'a self, 
        sp: S, 
        msg: &str
    ) -> DiagnosticBuilder<'a>;

    fn cannot_move_when_borrowed(
        &self, 
        span: Span, 
        desc: &str, 
        o: Origin
    ) -> DiagnosticBuilder { ... }

    fn cannot_use_when_mutably_borrowed(
        &self, 
        span: Span, 
        desc: &str, 
        borrow_span: Span, 
        borrow_desc: &str, 
        o: Origin
    ) -> DiagnosticBuilder { ... }

    fn cannot_act_on_uninitialized_variable(
        &self, 
        span: Span, 
        verb: &str, 
        desc: &str, 
        o: Origin
    ) -> DiagnosticBuilder { ... }

    fn cannot_mutably_borrow_multiply(
        &self, 
        new_loan_span: Span, 
        desc: &str, 
        opt_via: &str, 
        old_loan_span: Span, 
        old_opt_via: &str, 
        old_load_end_span: Span, 
        o: Origin
    ) -> DiagnosticBuilder { ... }

    fn cannot_uniquely_borrow_by_two_closures(
        &self, 
        new_loan_span: Span, 
        desc: &str, 
        old_loan_span: Span, 
        old_load_end_span: Span, 
        o: Origin
    ) -> DiagnosticBuilder { ... }

    fn cannot_uniquely_borrow_by_one_closure(
        &self, 
        new_loan_span: Span, 
        desc_new: &str, 
        opt_via: &str, 
        old_loan_span: Span, 
        noun_old: &str, 
        old_opt_via: &str, 
        previous_end_span: Span, 
        o: Origin
    ) -> DiagnosticBuilder { ... }

    fn cannot_reborrow_already_uniquely_borrowed(
        &self, 
        new_loan_span: Span, 
        desc_new: &str, 
        opt_via: &str, 
        kind_new: &str, 
        old_loan_span: Span, 
        old_opt_via: &str, 
        previous_end_span: Span, 
        o: Origin
    ) -> DiagnosticBuilder { ... }

    fn cannot_reborrow_already_borrowed(
        &self, 
        span: Span, 
        desc_new: &str, 
        msg_new: &str, 
        kind_new: &str, 
        old_span: Span, 
        noun_old: &str, 
        kind_old: &str, 
        msg_old: &str, 
        old_load_end_span: Span, 
        o: Origin
    ) -> DiagnosticBuilder { ... }

    fn cannot_assign_to_borrowed(
        &self, 
        span: Span, 
        borrow_span: Span, 
        desc: &str, 
        o: Origin
    ) -> DiagnosticBuilder { ... }

    fn cannot_move_into_closure(
        &self, 
        span: Span, 
        desc: &str, 
        o: Origin
    ) -> DiagnosticBuilder { ... }

    fn cannot_reassign_immutable(
        &self, 
        span: Span, 
        desc: &str, 
        o: Origin
    ) -> DiagnosticBuilder { ... }

    fn cannot_assign(
        &self, 
        span: Span, 
        desc: &str, 
        o: Origin
    ) -> DiagnosticBuilder { ... }

    fn cannot_assign_static(
        &self, 
        span: Span, 
        desc: &str, 
        o: Origin
    ) -> DiagnosticBuilder { ... }

    fn cannot_move_out_of(
        &self, 
        move_from_span: Span, 
        move_from_desc: &str, 
        o: Origin
    ) -> DiagnosticBuilder { ... }

    fn cannot_move_out_of_interior_noncopy(
        &self, 
        move_from_span: Span, 
        ty: Ty, 
        is_index: bool, 
        o: Origin
    ) -> DiagnosticBuilder { ... }

    fn cannot_move_out_of_interior_of_drop(
        &self, 
        move_from_span: Span, 
        container_ty: Ty, 
        o: Origin
    ) -> DiagnosticBuilder { ... }

    fn cannot_act_on_moved_value(
        &self, 
        use_span: Span, 
        verb: &str, 
        optional_adverb_for_moved: &str, 
        moved_path: &str, 
        o: Origin
    ) -> DiagnosticBuilder { ... }

    fn cannot_partially_reinit_an_uninit_struct(
        &self, 
        span: Span, 
        uninit_path: &str, 
        o: Origin
    ) -> DiagnosticBuilder { ... }

    fn closure_cannot_assign_to_borrowed(
        &self, 
        span: Span, 
        descr: &str, 
        o: Origin
    ) -> DiagnosticBuilder { ... }

    fn cannot_borrow_path_as_mutable(
        &self, 
        span: Span, 
        path: &str, 
        o: Origin
    ) -> DiagnosticBuilder { ... }

    fn cannot_borrow_across_generator_yield(
        &self, 
        span: Span, 
        yield_span: Span, 
        o: Origin
    ) -> DiagnosticBuilder { ... }

    fn path_does_not_live_long_enough(
        &self, 
        span: Span, 
        path: &str, 
        o: Origin
    ) -> DiagnosticBuilder { ... }

    fn lifetime_too_short_for_reborrow(
        &self, 
        span: Span, 
        path: &str, 
        o: Origin
    ) -> DiagnosticBuilder { ... }

    fn cannot_act_on_capture_in_sharable_fn(
        &self, 
        span: Span, 
        bad_thing: &str, 
        help: (Span, &str), 
        o: Origin
    ) -> DiagnosticBuilder { ... }

    fn cannot_assign_into_immutable_reference(
        &self, 
        span: Span, 
        bad_thing: &str, 
        o: Origin
    ) -> DiagnosticBuilder { ... }

    fn cannot_capture_in_long_lived_closure(
        &self, 
        closure_span: Span, 
        borrowed_path: &str, 
        capture_span: Span, 
        o: Origin
    ) -> DiagnosticBuilder { ... }
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Required Methods

fn struct_span_err_with_code<'a, S: Into<MultiSpan>>(
&'a self,
sp: S,
msg: &str,
code: &str
) -> DiagnosticBuilder<'a>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn struct_span_err<'a, S: Into<MultiSpan>>(
&'a self,
sp: S,
msg: &str
) -> DiagnosticBuilder<'a>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Provided Methods

fn cannot_move_when_borrowed(
&self,
span: Span,
desc: &str,
o: Origin
) -> DiagnosticBuilder

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn cannot_use_when_mutably_borrowed(
&self,
span: Span,
desc: &str,
borrow_span: Span,
borrow_desc: &str,
o: Origin
) -> DiagnosticBuilder

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn cannot_act_on_uninitialized_variable(
&self,
span: Span,
verb: &str,
desc: &str,
o: Origin
) -> DiagnosticBuilder

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn cannot_mutably_borrow_multiply(
&self,
new_loan_span: Span,
desc: &str,
opt_via: &str,
old_loan_span: Span,
old_opt_via: &str,
old_load_end_span: Span,
o: Origin
) -> DiagnosticBuilder

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn cannot_uniquely_borrow_by_two_closures(
&self,
new_loan_span: Span,
desc: &str,
old_loan_span: Span,
old_load_end_span: Span,
o: Origin
) -> DiagnosticBuilder

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn cannot_uniquely_borrow_by_one_closure(
&self,
new_loan_span: Span,
desc_new: &str,
opt_via: &str,
old_loan_span: Span,
noun_old: &str,
old_opt_via: &str,
previous_end_span: Span,
o: Origin
) -> DiagnosticBuilder

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn cannot_reborrow_already_uniquely_borrowed(
&self,
new_loan_span: Span,
desc_new: &str,
opt_via: &str,
kind_new: &str,
old_loan_span: Span,
old_opt_via: &str,
previous_end_span: Span,
o: Origin
) -> DiagnosticBuilder

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn cannot_reborrow_already_borrowed(
&self,
span: Span,
desc_new: &str,
msg_new: &str,
kind_new: &str,
old_span: Span,
noun_old: &str,
kind_old: &str,
msg_old: &str,
old_load_end_span: Span,
o: Origin
) -> DiagnosticBuilder

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn cannot_assign_to_borrowed(
&self,
span: Span,
borrow_span: Span,
desc: &str,
o: Origin
) -> DiagnosticBuilder

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn cannot_move_into_closure(
&self,
span: Span,
desc: &str,
o: Origin
) -> DiagnosticBuilder

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn cannot_reassign_immutable(
&self,
span: Span,
desc: &str,
o: Origin
) -> DiagnosticBuilder

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn cannot_assign(&self, span: Span, desc: &str, o: Origin) -> DiagnosticBuilder

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn cannot_assign_static(
&self,
span: Span,
desc: &str,
o: Origin
) -> DiagnosticBuilder

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn cannot_move_out_of(
&self,
move_from_span: Span,
move_from_desc: &str,
o: Origin
) -> DiagnosticBuilder

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn cannot_move_out_of_interior_noncopy(
&self,
move_from_span: Span,
ty: Ty,
is_index: bool,
o: Origin
) -> DiagnosticBuilder

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn cannot_move_out_of_interior_of_drop(
&self,
move_from_span: Span,
container_ty: Ty,
o: Origin
) -> DiagnosticBuilder

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn cannot_act_on_moved_value(
&self,
use_span: Span,
verb: &str,
optional_adverb_for_moved: &str,
moved_path: &str,
o: Origin
) -> DiagnosticBuilder

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn cannot_partially_reinit_an_uninit_struct(
&self,
span: Span,
uninit_path: &str,
o: Origin
) -> DiagnosticBuilder

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn closure_cannot_assign_to_borrowed(
&self,
span: Span,
descr: &str,
o: Origin
) -> DiagnosticBuilder

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn cannot_borrow_path_as_mutable(
&self,
span: Span,
path: &str,
o: Origin
) -> DiagnosticBuilder

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn cannot_borrow_across_generator_yield(
&self,
span: Span,
yield_span: Span,
o: Origin
) -> DiagnosticBuilder

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn path_does_not_live_long_enough(
&self,
span: Span,
path: &str,
o: Origin
) -> DiagnosticBuilder

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn lifetime_too_short_for_reborrow(
&self,
span: Span,
path: &str,
o: Origin
) -> DiagnosticBuilder

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn cannot_act_on_capture_in_sharable_fn(
&self,
span: Span,
bad_thing: &str,
help: (Span, &str),
o: Origin
) -> DiagnosticBuilder

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn cannot_assign_into_immutable_reference(
&self,
span: Span,
bad_thing: &str,
o: Origin
) -> DiagnosticBuilder

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn cannot_capture_in_long_lived_closure(
&self,
closure_span: Span,
borrowed_path: &str,
capture_span: Span,
o: Origin
) -> DiagnosticBuilder

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?