Function rustc_mir::util::dump_enabled

pub fn dump_enabled<'a, 'tcx>(
    tcx: TyCtxt<'a, 'tcx, 'tcx>, 
    pass_name: &str, 
    source: MirSource
) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?