Function rustc_mir::util::dump_mir

pub fn dump_mir<'a, 'tcx>(
    tcx: TyCtxt<'a, 'tcx, 'tcx>, 
    pass_num: Option<(MirSuite, MirPassIndex)>, 
    pass_name: &str, 
    disambiguator: &Display, 
    source: MirSource, 
    mir: &Mir<'tcx>
)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

If the session is properly configured, dumps a human-readable representation of the mir into:

rustc.node<node_id>.<pass_num>.<pass_name>.<disambiguator>

Output from this function is controlled by passing -Z dump-mir=<filter>, where <filter> takes the following forms:

all -- dump MIR for all fns, all passes, all everything

substring1&substring2,... -- &-separated list of substrings that can appear in the pass-name or the item_path_str for the given node-id. If any one of the substrings match, the data is dumped out.