Function rustc_mir::util::write_graphviz_node_label

pub fn write_graphviz_node_label<W: Write, INIT, FINI>(
    block: BasicBlock, 
    mir: &Mir, 
    w: &mut W, 
    num_cols: u32, 
    init: INIT, 
    fini: FINI
) -> Result<()> where
    INIT: Fn(&mut W) -> Result<()>,
    FINI: Fn(&mut W) -> Result<()>, 

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Write a graphviz HTML-styled label for the given basic block, with all necessary escaping already performed. (This is suitable for emitting directly, as is done in this module, or for use with the LabelText::HtmlStr from libgraphviz.)

init and fini are callbacks for emitting additional rows of data (using HTML enclosed with <tr> in the emitted text).