Function rustc_mir::util::write_mir_graphviz

pub fn write_mir_graphviz<'a, 'tcx, W>(
    tcx: TyCtxt<'a, 'tcx, 'tcx>, 
    single: Option<DefId>, 
    w: &mut W
) -> Result<()> where
    W: Write, 

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Write a graphviz DOT graph of a list of MIRs.